3-11 게시물 수정과 삭제 (answer 삭제 user 검증 추가)

+++ app/domain/pybo/answer/services/service.py
@@ -58,11 +58,10 @@ class AnswerService:
         반환값:
         - AnswerItemResponse: answer 하나가 포함된 성공 응답을 반환합니다.
         """
-        async with db.begin():
-            response = await self.answer_repository.find_answer(db, answer_id)
+        response = await self.answer_repository.find_answer(db, answer_id)
 
         if response is None:
-            raise HTTPException(status_code=HTTPStatus.NOT_FOUND, detail="Answer not found")
+            raise HTTPException(status_code=HTTPStatus.BAD_REQUEST, detail="Answer not found")
 
         return AnswerItemResponse.model_validate(response)
 
@@ -86,36 +85,37 @@ class AnswerService:
         return AnswerItemResponse.model_validate(response)
 
 
-    async def update_answer(self, db: AsyncSession, answer_id: int, update_dto: AnswerUpdateRequest) -> AnswerItemAffectedResponse:
+    async def update_answer(self, db: AsyncSession, answer_id: int, update_dto: AnswerUpdateRequest, user: UserItemResponse) -> AnswerItemAffectedResponse:
         """
         answer 하나를 수정하는 비동기 서비스
 
         매개변수:
         - db (AsyncSession): 비동기 데이터베이스 세션을 사용합니다.
         - answer_id (int): answer 하나의 고유 ID를 전달합니다.
-        - update_dto (AnswerUpdateRequest): answer 수정을 위한 데이터를 전달합니다.
+        - update_dto (AnswerUpdateRequest): answer 수정을 위한 폼 데이터를 전달합니다.
+        - user (UserItemResponse): 사용자의 정보를 포함하는 UserItemResponse를 전달합니다.
 
         반환값:
         - AnswerItemAffectedResponse: 수정된 answer의 개수가 포함된 성공 응답을 반환합니다.
         """
         async with db.begin():
+            answer = await self.find_answer(db=db, answer_id=answer_id)
+
+            if answer.user_id != user.id:
+                raise HTTPException(status_code=HTTPStatus.FORBIDDEN, detail="Forbidden user")
+
             response = await self.answer_repository.update_answer(db, answer_id, update_dto.model_dump())
 
         return AnswerItemAffectedResponse.model_validate({"rowcount": response})
 
 
-    async def delete_answer(self, db: AsyncSession, answer_id: int) -> AnswerItemAffectedResponse:
-        """
-        answer 하나를 삭제하는 비동기 서비스
-
-        매개변수:
-        - db (AsyncSession): 비동기 데이터베이스 세션을 사용합니다.
-        - answer_id (int): answer 하나의 고유 ID를 전달합니다.
-
-        반환값:
-        - AnswerItemAffectedResponse: 삭제된 answer의 개수가 포함된 성공 응답을 반환합니다.
-        """
+    async def delete_answer(self, db: AsyncSession, answer_id: int, user: UserItemResponse) -> AnswerItemAffectedResponse:
         async with db.begin():
+            answer = await self.find_answer(db, answer_id)
+
+            if answer.user_id != user.id:
+                raise HTTPException(status_code=HTTPStatus.FORBIDDEN, detail="Forbidden user")
+
             response = await self.answer_repository.delete_answer(db, answer_id)
 
         return AnswerItemAffectedResponse.model_validate({"rowcount": response})
